requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: win-64
# created-by: conda 24.11.3
annotated-types=0.7.0=pypi_0
anyio=4.9.0=pypi_0
bzip2=1.0.8=h2bbff1b_6
ca-certificates=2025.2.25=haa95532_0
certifi=2025.1.31=pypi_0
click=8.1.8=pypi_0
colorama=0.4.6=pypi_0
dnspython=2.7.0=pypi_0
email-validator=2.2.0=pypi_0
exceptiongroup=1.2.2=pypi_0
fastapi=0.115.12=pypi_0
fastapi-cli=0.0.7=pypi_0
h11=0.14.0=pypi_0
httpcore=1.0.7=pypi_0
httptools=0.6.4=pypi_0
httpx=0.28.1=pypi_0
idna=3.10=pypi_0
jinja2=3.1.6=pypi_0
libffi=3.4.4=hd77b12b_1
markdown-it-py=3.0.0=pypi_0
markupsafe=3.0.2=pypi_0
mdurl=0.1.2=pypi_0
motor=3.7.0=pypi_0
openssl=3.0.16=h3f729d1_0
pip=25.0=py310haa95532_0
pydantic=2.11.0=pypi_0
pydantic-core=2.33.0=pypi_0
pygments=2.19.1=pypi_0
pymongo=4.11.3=pypi_0
python=3.10.16=h4607a30_1
python-dotenv=1.1.0=pypi_0
python-multipart=0.0.20=pypi_0
pyyaml=6.0.2=pypi_0
rich=13.9.4=pypi_0
rich-toolkit=0.14.0=pypi_0
setuptools=75.8.0=py310haa95532_0
shellingham=1.5.4=pypi_0
sniffio=1.3.1=pypi_0
sqlite=3.45.3=h2bbff1b_0
starlette=0.46.1=pypi_0
tk=8.6.14=h0416ee5_0
typer=0.15.2=pypi_0
typing-extensions=4.13.0=pypi_0
typing-inspection=0.4.0=pypi_0
tzdata=2025a=h04d1e81_0
uvicorn=0.34.0=pypi_0
vc=14.42=haa95532_4
vs2015_runtime=14.42.34433=he0abc0d_4
watchfiles=1.0.4=pypi_0
websockets=15.0.1=pypi_0
wheel=0.45.1=py310haa95532_0
xz=5.6.4=h4754444_1
zlib=1.2.13=h8cc25b3_1
